testing proliferation of handlers
